+++ node/constants/index.ts
@@ -1,3 +1,7 @@
 export const APP_NAME = 'vtex.cross-device-cart'
 export const BUCKET = 'cross_device_cart'
 export const SETTINGS_PATH = 'app_settings'
+export const DEFAULT_SETTINGS = {
+  isAutomatic: true,
+  strategy: 'REPLACE',
+} as const