+++ node/resolvers/index.ts
@@ -2,7 +2,7 @@ import { getAppSettings } from './getAppSettings'
 import { getSavedCart } from './getSavedCart'
 import { saveAppSettings } from './saveAppSettings'
 import { saveCurrentCart } from './saveCurrentCart'
-import { mergeCarts } from './mergeCarts'
+import { replaceCart } from './replaceCart'
 
 export const queries = {
   getAppSettings,
@@ -12,5 +12,5 @@ export const queries = {
 export const mutations = {
   saveAppSettings,
   saveCurrentCart,
-  mergeCarts,
+  replaceCart,
 }