+++ react/components/ChallengeBlock.tsx
@@ -1,71 +1,31 @@
-import React, { FC, Fragment, useState } from 'react'
+import React, { FC, Fragment } from 'react'
 import { useDevice } from 'vtex.device-detector'
 import { FormattedMessage } from 'react-intl'
 import { Button, ButtonWithIcon, IconClose } from 'vtex.styleguide'
 import { useCssHandles } from 'vtex.css-handles'
-
-import { MergeOptionsModal } from './MergeOptionsModal'
 
 const CSS_HANDLES = ['actionBar', 'challengeText'] as const
 
 const closeIcon = <IconClose />
 
 interface Props {
-  handleAccept: (
-    showToast: (toast: ToastParam) => void,
-    strategy: MergeStrategy
-  ) => Promise<void>
-  toastHandler: (toast: ToastParam) => void
+  handleAccept: () => void
   handleDecline: () => void
-  mergeStrategy: MergeStrategy
   mutationLoading: boolean
-  advancedOptions: boolean
-  /* items: unknown[] */
 }
 
 const ChallengeBlock: FC<Props> = ({
-  mergeStrategy,
   handleAccept,
   handleDecline,
   mutationLoading,
-  toastHandler,
-  advancedOptions,
 }) => {
   const { device } = useDevice()
-  const [isModalOpen, setIsModalOpen] = useState(false)
   const handles = useCssHandles(CSS_HANDLES)
-
-  const handleCloseModal = () => {
-    setIsModalOpen(false)
-  }
 
   const deviceClass = `${device === 'phone' ? 'flex-column' : ''}`
 
-  const callToAction = (
-    <Button
-      size="small"
-      variation="secondary"
-      onClick={() => {
-        advancedOptions
-          ? setIsModalOpen(true)
-          : handleAccept(toastHandler, mergeStrategy)
-      }}
-      isLoading={mutationLoading}
-    >
-      <FormattedMessage id="store/crossCart.challenge.cta" />
-    </Button>
-  )
-
   return (
     <Fragment>
-      <MergeOptionsModal
-        /* items={items} */
-        strategies={['ADD', 'COMBINE', 'REPLACE']}
-        isOpen={isModalOpen}
-        handleClose={handleCloseModal}
-        handleAccept={handleAccept}
-        toastHandler={toastHandler}
-      />
       <div
         className={`${deviceClass} pa4 tc w-100 bg-base flex items-center justify-center ${handles.actionBar}`}
       >
@@ -77,7 +37,18 @@ const ChallengeBlock: FC<Props> = ({
           <FormattedMessage id="store/crossCart.challenge.text" />
         </span>
         <div className="flex">
-          <span className="mh4">{callToAction}</span>
+          <span className="mh4">
+            <Button
+              size="small"
+              variation="secondary"
+              onClick={() => {
+                handleAccept()
+              }}
+              isLoading={mutationLoading}
+            >
+              <FormattedMessage id="store/crossCart.challenge.cta" />
+            </Button>
+          </span>
           <span>
             <ButtonWithIcon
               size="small"