+++ CHANGELOG.md
@@ -7,7 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2022-02-26
+
+## [1.0.1] - 2022-02-23
+
+- Fixed settings on first time installation
+
+## [1.0.0] - 2022-02-23
+
+- New version replacing how the cross cart is handled
+
 ## [0.1.0] - 2022-02-17
 
 ### Added
+
 - Initial release.